8.1 Correct statistics display
